% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{This is example data to be included in my package}
\description{
This is example data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
Rob Eva \email{robeva5921@gmail.com}
}
\keyword{data}
