% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_function}
\alias{my_function}
\title{Example Function to multiply two numbers}
\usage{
my_function(x, y)
}
\arguments{
\item{x}{one of the arguments to multiply.}

\item{y}{one of the arguments to multiply.}
}
\description{
This is a simple function that, by default, multiplies two numbers.
}
