% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{overdue}
\alias{overdue}
\title{Main Function}
\usage{
overdue(ch_data, mode = "")
}
\arguments{
\item{ch_data}{the dataset to work on.}

\item{mode}{the type of account to include - default is all.}
}
\description{
This is calculates the percentage of overdue accounts.
}
